fix: process cleanup
