+++ gpustack/worker/rpc_server.py
@@ -8,7 +8,7 @@ import sys
 import setproctitle
 
 from gpustack.utils.compat_importlib import pkg_resources
-from gpustack.utils.signal import add_signal_handlers
+from gpustack.utils.process import add_signal_handlers
 from gpustack.worker.backends.base import get_env_name_by_vendor
 from gpustack.worker.backends.llama_box import get_llama_box_command
